This is first prototype of the project
This prototype can only work in google colab and not in your local machine. This is only a text to image generator and does not fulfill the main project idea. To fulfil the main project idea we need to make a small project which can generate images so this fulfills that purpose.