This is the second prototype of the project.
This prototype solves the problem of running in the local machine. So, the first prototype was not able to run in the local machine which can be a laptop for example. But this resolves that issue. Also, this is also a text to image generator and does not fulfills the main purpose of the project.